adding EngineTests and moving out of NodeTests
